% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripe_functions.R
\name{get_all_plans}
\alias{get_all_plans}
\title{Get All Plans from Stripe API}
\usage{
get_all_plans(secret_key)
}
\arguments{
\item{secret_key}{A string containing the secret key for Stripe API authentication.}
}
\value{
An object of class \code{response} representing the API response.
}
\description{
This function retrieves all the plans from the Stripe API.
It requires a secret key for authentication.
}
