% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripe_functions.R
\name{get_customer_by_email}
\alias{get_customer_by_email}
\title{Retrieve Customer Data by Email from Stripe API}
\usage{
get_customer_by_email(email, secret_key)
}
\arguments{
\item{email}{The email address of the customer to be retrieved.}

\item{secret_key}{A string containing the secret key for Stripe API authentication.}
}
\value{
An object of class \code{response} representing the API response, which contains
the customer data if found, or an error message otherwise.
}
\description{
This function fetches customer data from the Stripe API based on the provided email address.
It requires an email address and a secret key for authentication.
}
