% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripe_functions.R
\name{get_customer_sub_by_id}
\alias{get_customer_sub_by_id}
\title{Retrieve Customer Subscription Information by ID from Stripe API}
\usage{
get_customer_sub_by_id(secret_key, customer_id)
}
\arguments{
\item{secret_key}{A string containing the secret key for Stripe API authentication.}

\item{customer_id}{The unique identifier of the customer whose subscription information is to be retrieved.}
}
\value{
An object of class \code{response} representing the API response, which contains
the subscription information for the given customer ID.
}
\description{
This function fetches subscription information for a specific customer from the Stripe API.
It requires a customer ID and a secret key for authentication.
}
