% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripe_functions.R
\name{get_user}
\alias{get_user}
\title{Retrieve Customer and Subscription Data from Stripe}
\usage{
get_user(email, secret_key)
}
\arguments{
\item{email}{A string representing the customer's email.}

\item{secret_key}{A string representing the Stripe secret key for authentication.}
}
\value{
A list containing two elements: \code{res_sub} (subscription data) and \code{res_cus} (customer data).
Returns NULL if no data is found or in case of an error.
}
\description{
This function fetches customer data based on the email provided and retrieves associated subscription data using the Stripe API.
}
\examples{
\dontrun{
  user_data <- get_user("customer@example.com", "your_secret_key")
  print(user_data)
}

}
