% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripe_functions.R
\name{has_plan}
\alias{has_plan}
\title{Check Customer's Subscription Status}
\usage{
has_plan(email, secret_key)
}
\arguments{
\item{email}{A string representing the customer's email.}

\item{secret_key}{A string representing the Stripe secret key for authentication.}
}
\value{
A logical value: TRUE if the customer has an active subscription, FALSE otherwise.
}
\description{
This function checks whether a given customer, identified by their email, has any active subscription in Stripe.
}
\examples{
\dontrun{
  subscription_status <- has_plan("customer@example.com", "your_secret_key")
  print(subscription_status)
}

}
