% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripe_functions.R
\name{make_api_request}
\alias{make_api_request}
\title{Make an API Request to Stripe}
\usage{
make_api_request(
  endpoint,
  query_list,
  secret_key,
  api_base_url = "https://api.stripe.com/v1"
)
}
\arguments{
\item{endpoint}{A string specifying the API endpoint to be accessed.}

\item{query_list}{A list of query parameters to be included in the request.}

\item{secret_key}{A string representing the Stripe secret key for authentication.}

\item{api_base_url}{A string specifying the base URL of the Stripe API. Default is 'https://api.stripe.com/v1'.}
}
\value{
A response object from the Stripe API, or NULL in case of an error.
}
\description{
This function makes a GET request to a specified Stripe API endpoint using provided query parameters and authentication details.
It is a helper function to interact with the Stripe API.
}
\examples{
\dontrun{
  response <- make_api_request("/customers",
                              list(email = "customer@example.com"),
                              "your_secret_key")
}

}
