% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripe_functions.R
\name{user_has_plan_prod}
\alias{user_has_plan_prod}
\title{Check if User Has an Active or Trialing Plan Based on Email}
\usage{
user_has_plan_prod(email, secret_key)
}
\arguments{
\item{email}{The email address of the user.}

\item{secret_key}{A string containing the secret key for Stripe API authentication.}
}
\value{
A list containing a boolean field \code{has_plan} indicating whether the user has an
active or trialing plan, and \code{content_sub_prod} which is the product associated
with the subscription. Returns \code{FALSE} if the user does not have an active or
trialing plan, or if the customer ID is not found.
}
\description{
This function checks if a user, identified by their email, has an active or trialing
subscription plan in Stripe. It first retrieves the customer's ID based on their email,
then fetches their subscription details to determine the status and product of the subscription.
}
